Pass filesizes to server as i64.
<==== Mercurial data ====>
Branch: vms_3.1_wearable

+++ common/src/api/server_rest_connection.cpp
@@ -267,7 +267,7 @@ Handle ServerConnection::checkCloudHost(
 
 Handle ServerConnection::addFileDownload(
     const QString& fileName,
-    int size,
+    qint64 size,
     const QByteArray& md5,
     const QUrl& url,
     GetCallback callback,
@@ -287,8 +287,8 @@ Handle ServerConnection::addFileDownload(
 
 Handle ServerConnection::addFileUpload(
     const QString& fileName,
-    int size,
-    int chunkSize,
+    qint64 size,
+    qint64 chunkSize,
     const QByteArray& md5,
     PostCallback callback,
     QThread* targetThread)
